requirements.txt
fastapi
uvicorn
python-multipart
streamlit
opencv-python-headless
numpy
ultralytics
llama-cpp-python
requests
pydantic